refactor: Car 클래스를 분리하여 코드 구조 개선
